Small refactor

+++ db/migrate/20230724152517_delete_all_slovensko_sk_draft_messages.rb
@@ -2,12 +2,8 @@ class DeleteAllSlovenskoSkDraftMessages < ActiveRecord::Migration[7.0]
   def change
     Message.joins(:tags).where(tags: { name: 'slovensko.sk:Drafts' }).find_each do |message|
       thread = message.thread
-      if thread.messages.count == 1
-        message.destroy
-        thread.destroy
-      else
-        message.destroy
-      end
+      message.destroy
+      thread.destroy unless thread.messages.any?
     end
   end
 end
